# swagger.yaml
openapi: 3.0.3
info:
  title: Book Rating System API
  version: 1.0.0
  description: >
    A secure REST API for managing books and ratings with role-based access control.
    Supports JWT authentication via HttpOnly cookies, RBAC (admin/reader),
    pagination, search, and Redis caching.
  contact:
    name: BookShelf Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in an HttpOnly secure cookie
  schemas:
...
    UpdateBookRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        description:
          type: string
        genre:
          type: string
        published_year:
          type: integer
...
    UpdateRatingRequest:
      type: object
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        review:
          type: string
          example: Updated — even better on a second reading!

    # ── Response Bodies ─────────────────────────────────────────────
...
    Book:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: The Great Gatsby
        author:
          type: string
          example: F. Scott Fitzgerald
        isbn:
          type: string
          example: "978-0743273565"
        description:
          type: string
        genre:
          type: string
          example: Fiction
        published_year:
          type: integer
          example: 1925
        avg_rating:
          type: number
          format: float
          example: 4.3
        created_at:
          type: string
          format: date-time
    Rating:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 2
        book_id:
          type: integer
          example: 1
        score:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        review:
          type: string
          example: A beautifully written classic.
        username:
          type: string
          example: johndoe
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation error
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
paths:
...
  /books/{id}:
    get:
      tags: [Books]
      summary: Get a single book
      description: Returns a single book by ID, including its average rating.
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Book ID
      responses:
        "200":
          description: Book details with average rating
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Book"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags: [Books]
      summary: Update a book
      description: "**Admin only.** Updates one or more fields of an existing book."
      operationId: updateBook
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Book ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookRequest"
      responses:
        "200":
          description: Book updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Book updated successfully
                  data:
                    $ref: "#/components/schemas/Book"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not authorized — admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Books]
      summary: Delete a book
      description: "**Admin only.** Permanently removes a book and its associated ratings."
      operationId: deleteBook
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Book ID
      responses:
        "200":
          description: Book deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Book deleted successfully
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not authorized — admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Ratings ─────────────────────────────────────────────────────
...
  /books/{bookId}/ratings/{id}:
    put:
      tags: [Ratings]
      summary: Update a rating
      description: "**Owner only.** Updates the authenticated user's own rating."
      operationId: updateRating
      security:
        - cookieAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
          description: Book ID
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Rating ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRatingRequest"
      responses:
        "200":
          description: Rating updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Rating updated successfully
                  data:
                    $ref: "#/components/schemas/Rating"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not authorized — must be rating owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Rating not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Ratings]
      summary: Delete a rating
      description: "**Owner or Admin.** Removes a rating permanently."
      operationId: deleteRating
      security:
        - cookieAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
          description: Book ID
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Rating ID
      responses:
        "200":
          description: Rating deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Rating deleted successfully
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not authorized — must be owner or admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Rating not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"